+++ src/js/Manager/constants.js
