+++ src/js/Manager/validation.js
