+++ src/js/app.js
@@ -1,7 +1,7 @@
 
 import Storage from './Storage';
-import SearchVideoManager from './SearchVideoManager';
-import SaveVideoManager from './SaveVideoManager';
+import SearchVideoManager from './Manager/SearchVideoManager';
+import SaveVideoManager from './Manager/SaveVideoManager';
 import HomeView from './View/HomeView';
 import SearchModalView from './View/SearchModalView';
